Allow extend on same size.
The reason for this is that OS decides the size of the volume before
calling the driver and before a potential resize operation comes
to the driver.

In the case of volume create from cache image, the cached image has
the orignial size, but the created volume already has a bigger
(extended) size before the call is made to the driver API.

Thus, the resize operation was blocked by the driver, which was
assuming is a no-op.

With this change, a same size resize is accepted, which solves the
use case above. Even if the Datera API is called with the same size,
the 'resize to the same size' operation is a no op on the Datera
side, so that case is covered as well.